ramips: Add support for Wavlink WL-WN575A2
This commit adds support for the Wavlink WL-WN575A2, a dual-band wall-plug
wireless router with the following specifications:

- CPU: MediaTek MT7628AN 580MHz
 - Flash: 8MB
 - RAM: 64MB
 - Ethernet: 2x 10/100 Mbps (switched)
 - 2.4 GHz: 802.11b/g/n SoC, MIMO 2x2, 20 dBm
 - 5 GHz: 802.11a/n/ac MT7610E, 1x1, 20dBm
 - Antennas: 3x external (2x2.4GHz and 1x5GHz)
 - LEDs: 4 programmable + LAN, WAN, POWER
 - Buttons: reset, WPS

It's the same board of Wavlink WL-WN575A3 except for the 5GHz chip (MT7610E vs MT7612E).

Everything is working except for the 5GHz wireless, because is not supported
by the mt76 package.

Flashing instructions:

Factory U-boot launches a TFTP client if WPS button is pressed during power-on.
Rename the sysupgrade file and configure a TFTP as follows:

 - Client (WL-WN575A2) IP: 192.168.10.101
 - Server IP: 192.168.10.100
 - Filename: firmware.bin

Via serial port (there is a very tiny T/R next to the bios chip) is possible
to stop the boot process. At this point is possible to choose one of these options:
   1: Load system code to SDRAM via TFTP.
   2: Load system code then write to Flash via TFTP.
   3: Boot system code via Flash (default).
   4: Entr boot command line interface.
   7: Load Boot Loader code then write to Flash via Serial.
   9: Load Boot Loader code then write to Flash via TFTP.

By pressing 1 is possible to load an initramfs image and try the firmware without
removing the original firmare

Serial port settings: 57600/8-N-1

Signed-off-by: Davide Fioravanti <pantanastyle@gmail.com>